Move ClientFileExclusions to Java client utils
